# files/files-api-codegen/files-java-dto-definitions/target/generated-sources/swagger/src/files/v4/r0/a1/file_servers/EnforcementType.java
/*
 * Generated file ..  Product version: 1.0.0-SNAPSHOT
 *
 * Part of the This is the files AP definitions
 *
 * (c) 2019 Nutanix Inc.  All rights reserved
 *
 */

package files.v4.r0.a1.file_servers;


import javax.validation.constraints.*;


public enum EnforcementType {

    HARD,
    SOFT,
    $UNKNOWN,

}